requirements.txt
iptools
dnspython
requests
lxml
pyyaml
straight.plugin
gevent
mmh3
